fastapi/project/requirements.txt
fastapi[standard]>=0.110.0
sqlalchemy>=2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic[email]>=2.0
uvicorn>=0.27.0
